onDeviceEvalAdapter is public and should not need external gson (#422)
### TL;DR

Don't require consumers of `OnDeviceEvalAdapter` to pass in a Gson instance

### Why make this change?

While I usually prefer dependency/constructor injection (which is why I was ok merging in the external PR), in this case we shouldn't change the signature of an externally-consumed class.
Since `StatsigUtil` now keeps a static `Gson` instance, this still maintains the performance benefits.
Also this change fixes the Kong breakage.

### What changed?

- replaced parameter with a call to `StatsigUtil.getOrBuildGson()`